@@ -33,15 +33,16 @@
     "ansi_up": "=1.3.0",
     "moment": "=2.13.0",
     "moment-duration-format": "=1.3.0",
-    "react": "=0.14.5",
-    "react-native": "=0.24.1",
-    "react-native-vector-icons": "=1.3.4",
+    "react": "=15.0.2",
+    "react-native": "=0.26.1",
+    "react-native-vector-icons": "=2.0.2",
     "underscore": "=1.8.3"
   },
   "devDependencies": {
-    "babel-eslint": "=6.0.3",
-    "eslint": "=2.8.0",
-    "eslint-plugin-react": "=5.0.1",
-    "eslint-plugin-react-native": "=1.0.0"
+    "babel-eslint": "=6.0.4",
+    "eslint": "=2.10.2",
+    "eslint-plugin-react": "=5.1.1",
+    "eslint-plugin-react-native": "=1.0.2",
+    "rnpm-plugin-upgrade": "=0.26.0"
   }
 }
