@@ -1,13 +1,13 @@
-import React from 'react-native';
+import React, { Component } from 'react'; // eslint-disable-line no-unused-vars
 import Icon from 'react-native-vector-icons/Octicons';
 
 import Constants from '../Utils/Constants';
 
-var {
+import {
   StyleSheet,
   Text,
   View,
-} = React;
+} from 'react-native';
 
 var styles = StyleSheet.create({
   container: {
@@ -40,7 +40,7 @@ var styles = StyleSheet.create({
   }
 });
 
-export default class StatusSidebar extends React.Component {
+export default class StatusSidebar extends Component {
   constructor(props) {
     super(props);
 
