@@ -1,9 +1,9 @@
-import React from 'react-native';
+import React, { Component } from 'react'; // eslint-disable-line no-unused-vars
 
-var {
+import {
   StyleSheet,
   View,
-} = React;
+} from 'react-native';
 
 var styles = StyleSheet.create({
   container: {
@@ -13,7 +13,7 @@ var styles = StyleSheet.create({
   }
 });
 
-export default class Separator extends React.Component {
+export default class Separator extends Component {
   render() {
     return (
       <View style={styles.container} />
