@@ -1,10 +1,10 @@
-import React from 'react-native';
+import React, { Component } from 'react'; // eslint-disable-line no-unused-vars
 
-var {
+import {
   StyleSheet,
   Text,
   View,
-} = React;
+} from 'react-native';
 
 var styles = StyleSheet.create({
   container: {
@@ -18,7 +18,7 @@ var styles = StyleSheet.create({
   }
 });
 
-export default class NavigationTitle extends React.Component {
+export default class NavigationTitle extends Component {
   render() {
     return (
       <View style={styles.container}>
