@@ -1,5 +1,5 @@
 import _ from 'underscore';
-import React from 'react-native';
+import React, { Component } from 'react'; // eslint-disable-line no-unused-vars
 
 import Api from '../Utils/Api';
 import EmptyResults from '../Components/EmptyResults';
@@ -7,12 +7,12 @@ import Loading from '../Components/Loading';
 import RepoItem from '../Components/RepoItem';
 import CustomRefreshControl from '../Helpers/CustomRefreshControl';
 
-var {
+import {
   ListView,
   ScrollView,
   StyleSheet,
   View
-} = React;
+} from 'react-native';
 
 var styles = StyleSheet.create({
   container: {
@@ -29,7 +29,7 @@ var styles = StyleSheet.create({
   }
 });
 
-export default class LatestProRepos extends React.Component {
+export default class LatestProRepos extends Component {
 
   constructor(props) {
     super(props);
